newly organized builder for linear chains
