+++ structure_tool/geometrical_functions.py
@@ -1,0 +1,34 @@
+import random
+import math
+import argparse
+import itertools
+import numpy as np
+import scipy.optimize as opt
+from numpy import sqrt, pi, cos, sin, dot, cross, arccos, degrees
+from numpy import float as nfl
+from numpy.linalg import norm
+
+def u_vect(vect):
+    return(vect/norm(vect))
+
+def angle(A, B, C):
+    v1 = B - A
+    v2 = B - C
+    return(degrees(arccos(np.clip(dot(u_vect(v1), u_vect(v2)), -1.0, 1.0))))
+
+def dih(A, B, C, D):
+    r1 = A - B
+    r2 = B - C
+    r3 = C - D
+    n1 = cross(r1, r2)
+    n2 = cross(r2, r3)
+    return(degrees(arccos(np.clip(dot(u_vect(n1), u_vect(n2)), -1.0, 1.0))))
+
+
+def geometrical_center(coord):
+    return(sum(coord)/float(len(coord)))
+
+def norm_sphere():
+    v_sphere = np.random.normal(0.0, 1, (5000,3))
+    return(np.array([ u_vect(vect) for vect in v_sphere]))
+
