--- structure_tool/out.gro
@@ -1,103 +1,0 @@
-Monte Carlo generated PS in THF
-   100     
-    1STYR     B    1   0.330   0.000   0.000
-    1STYR     R    2  -0.072   0.317  -0.054
-    1STYR     R    3  -0.122   0.609  -0.201
-    1STYR     R    4  -0.059   0.457  -0.488
-    2STYR     B    5  -0.298   0.485  -0.713
-    2STYR     R    6  -0.228   0.380  -1.018
-    2STYR     R    7  -0.288   0.686  -1.127
-    2STYR     R    8  -0.537   0.895  -1.071
-    3STYR     B    9  -0.381   0.993  -0.797
-    3STYR     R   10  -0.597   1.242  -0.786
-    3STYR     R   11  -0.666   1.480  -0.568
-    3STYR     R   12  -0.989   1.414  -0.550
-    4STYR     B   13  -1.243   1.562  -0.400
-    4STYR     R   14  -1.530   1.616  -0.554
-    4STYR     R   15  -1.516   1.833  -0.803
-    4STYR     R   16  -1.562   2.142  -0.909
-    5STYR     B   17  -1.575   2.469  -0.944
-    5STYR     R   18  -1.311   2.602  -0.796
-    5STYR     R   19  -1.034   2.627  -0.973
-    5STYR     R   20  -0.954   2.947  -0.957
-    6STYR     B   21  -0.869   3.107  -0.682
-    6STYR     R   22  -0.716   3.391  -0.612
-    6STYR     R   23  -0.856   3.401  -0.313
-    6STYR     R   24  -0.989   3.260  -0.045
-    7STYR     B   25  -1.105   2.965   0.044
-    7STYR     R   26  -1.117   2.916   0.370
-    7STYR     R   27  -1.178   2.622   0.506
-    7STYR     R   28  -1.393   2.546   0.267
-    8STYR     B   29  -1.629   2.349   0.387
-    8STYR     R   30  -1.814   2.265   0.128
-    8STYR     R   31  -1.657   2.179  -0.150
-    8STYR     R   32  -1.397   1.977  -0.162
-    9STYR     B   33  -1.169   1.829   0.025
-    9STYR     R   34  -0.880   1.989   0.019
-    9STYR     R   35  -0.673   1.858   0.240
-    9STYR     R   36  -0.665   1.727   0.542
-   10STYR     B   37  -0.410   1.782   0.745
-   10STYR     R   38  -0.509   1.773   1.060
-   10STYR     R   39  -0.244   1.725   1.250
-   10STYR     R   40  -0.250   1.447   1.427
-   11STYR     B   41   0.052   1.437   1.558
-   11STYR     R   42  -0.066   1.576   1.832
-   11STYR     R   43   0.063   1.789   2.049
-   11STYR     R   44   0.007   1.687   2.357
-   12STYR     B   45   0.145   1.586   2.640
-   12STYR     R   46  -0.088   1.375   2.741
-   12STYR     R   47  -0.026   1.257   3.043
-   12STYR     R   48  -0.324   1.231   3.183
-   13STYR     B   49  -0.389   1.264   3.505
-   13STYR     R   50  -0.447   1.577   3.591
-   13STYR     R   51  -0.582   1.798   3.796
-   13STYR     R   52  -0.508   1.847   4.114
-   14STYR     B   53  -0.283   1.996   4.305
-   14STYR     R   54  -0.472   2.253   4.390
-   14STYR     R   55  -0.408   2.184   4.706
-   14STYR     R   56  -0.567   2.105   4.984
-   15STYR     B   57  -0.534   2.036   5.305
-   15STYR     R   58  -0.253   1.939   5.448
-   15STYR     R   59  -0.287   1.823   5.755
-   15STYR     R   60  -0.269   1.920   6.070
-   16STYR     B   61  -0.531   1.773   6.207
-   16STYR     R   62  -0.763   1.848   6.429
-   16STYR     R   63  -1.093   1.840   6.441
-   16STYR     R   64  -1.179   1.846   6.122
-   17STYR     B   65  -1.329   2.128   6.038
-   17STYR     R   66  -1.550   2.146   5.794
-   17STYR     R   67  -1.549   2.165   5.464
-   17STYR     R   68  -1.331   2.412   5.487
-   18STYR     B   69  -1.013   2.495   5.505
-   18STYR     R   70  -0.800   2.746   5.484
-   18STYR     R   71  -0.539   2.948   5.499
-   18STYR     R   72  -0.382   2.963   5.789
-   19STYR     B   73  -0.442   2.663   5.911
-   19STYR     R   74  -0.140   2.555   5.835
-   19STYR     R   75  -0.027   2.288   5.677
-   19STYR     R   76   0.273   2.159   5.630
-   20STYR     B   77   0.538   1.963   5.608
-   20STYR     R   78   0.616   1.782   5.344
-   20STYR     R   79   0.938   1.728   5.392
-   20STYR     R   80   0.969   1.420   5.276
-   21STYR     B   81   0.934   1.122   5.413
-   21STYR     R   82   0.663   1.156   5.599
-   21STYR     R   83   0.444   1.342   5.761
-   21STYR     R   84   0.189   1.480   5.603
-   22STYR     B   85   0.069   1.212   5.451
-   22STYR     R   86   0.036   1.046   5.168
-   22STYR     R   87   0.117   0.783   4.987
-   22STYR     R   88  -0.058   0.688   4.724
-   23STYR     B   89  -0.176   0.937   4.543
-   23STYR     R   90  -0.182   1.067   4.240
-   23STYR     R   91  -0.273   0.995   3.931
-   23STYR     R   92  -0.241   0.764   3.698
-   24STYR     B   93  -0.461   0.676   3.468
-   24STYR     R   94  -0.783   0.619   3.511
-   24STYR     R   95  -0.991   0.667   3.763
-   24STYR     R   96  -1.133   0.721   4.056
-   25STYR     B   97  -1.225   1.013   4.178
-   25STYR     R   98  -1.340   1.293   4.044
-   25STYR     R   99  -1.318   1.271   3.716
-   25STYR     R  100  -1.216   1.131   3.435
-  2.00000 2.00000 2.00000
