+++ tests/structure_tool/100.itp
@@ -1,5 +1,5 @@
 [ nonbond_params ]
-PEO  PEO  0.43  3.6
+PEO  PEO  0.43 2.55 
 [   moleculetype   ]
 PEO   1
 [      atoms       ]
