+++ tests/structure_tool/20_nb.itp
@@ -1,0 +1,133 @@
+[ nonbond_params ]
+PEO  PEO  0.43  3.6
+[   moleculetype   ]
+PEO   1
+[      atoms       ]
+1     PEO   1 PEG   O1  1 0.000    45 
+2     PEO   1 PEG   O1  2 0.000    45 
+3     PEO   1 PEG   O1  3 0.000    45 
+4     PEO   1 PEG   O1  4 0.000    45 
+5     PEO   1 PEG   O1  5 0.000    45 
+6     PEO   1 PEG   O1  6 0.000    45 
+7     PEO   1 PEG   O1  7 0.000    45 
+8     PEO   1 PEG   O1  8 0.000    45 
+9     PEO   1 PEG   O1  9 0.000    45 
+10    PEO   1 PEG   O1  10 0.000    45 
+11    PEO   1 PEG   O1  11 0.000    45 
+12    PEO   1 PEG   O1  12 0.000    45 
+13    PEO   1 PEG   O1  13 0.000    45 
+14    PEO   1 PEG   O1  14 0.000    45 
+15    PEO   1 PEG   O1  15 0.000    45 
+16    PEO   1 PEG   O1  16 0.000    45 
+17    PEO   1 PEG   O1  17 0.000    45 
+18    PEO   1 PEG   O1  18 0.000    45 
+19    PEO   1 PEG   O1  19 0.000    45 
+20    PEO   1 PEG   O1  20 0.000    45 
+[      bonds       ]
+1     2     1  0.330    17000   
+2     3     1  0.330    17000   
+3     4     1  0.330    17000   
+4     5     1  0.330    17000   
+5     6     1  0.330    17000   
+6     7     1  0.330    17000   
+7     8     1  0.330    17000   
+8     9     1  0.330    17000   
+9     10    1  0.330    17000   
+10    11    1  0.330    17000   
+11    12    1  0.330    17000   
+12    13    1  0.330    17000   
+13    14    1  0.330    17000   
+14    15    1  0.330    17000   
+15    16    1  0.330    17000   
+16    17    1  0.330    17000   
+17    18    1  0.330    17000   
+18    19    1  0.330    17000   
+19    20    1  0.330    17000   
+[      angles      ]
+1     2     3     1  130.00   50      
+2     3     4     1  130.00   50      
+3     4     5     1  130.00   50      
+4     5     6     1  130.00   50      
+5     6     7     1  130.00   50      
+6     7     8     1  130.00   50      
+7     8     9     1  130.00   50      
+8     9     10    1  130.00   50      
+9     10    11    1  130.00   50      
+10    11    12    1  130.00   50      
+11    12    13    1  130.00   50      
+12    13    14    1  130.00   50      
+13    14    15    1  130.00   50      
+14    15    16    1  130.00   50      
+15    16    17    1  130.00   50      
+16    17    18    1  130.00   50      
+17    18    19    1  130.00   50      
+18    19    20    1  130.00   50      
+[    dihedrals     ]
+1     2     3     4     1  0          0.12       4         
+1     2     3     4     1  0          0.18       2         
+1     2     3     4     1  0          0.33       3         
+1     2     3     4     1  180.00     1.96       1         
+2     3     4     5     1  0          0.12       4         
+2     3     4     5     1  0          0.18       2         
+2     3     4     5     1  0          0.33       3         
+2     3     4     5     1  180.00     1.96       1         
+3     4     5     6     1  0          0.12       4         
+3     4     5     6     1  0          0.18       2         
+3     4     5     6     1  0          0.33       3         
+3     4     5     6     1  180.00     1.96       1         
+4     5     6     7     1  0          0.12       4         
+4     5     6     7     1  0          0.18       2         
+4     5     6     7     1  0          0.33       3         
+4     5     6     7     1  180.00     1.96       1         
+5     6     7     8     1  0          0.12       4         
+5     6     7     8     1  0          0.18       2         
+5     6     7     8     1  0          0.33       3         
+5     6     7     8     1  180.00     1.96       1         
+6     7     8     9     1  0          0.12       4         
+6     7     8     9     1  0          0.18       2         
+6     7     8     9     1  0          0.33       3         
+6     7     8     9     1  180.00     1.96       1         
+7     8     9     10    1  0          0.12       4         
+7     8     9     10    1  0          0.18       2         
+7     8     9     10    1  0          0.33       3         
+7     8     9     10    1  180.00     1.96       1         
+8     9     10    11    1  0          0.12       4         
+8     9     10    11    1  0          0.18       2         
+8     9     10    11    1  0          0.33       3         
+8     9     10    11    1  180.00     1.96       1         
+9     10    11    12    1  0          0.12       4         
+9     10    11    12    1  0          0.18       2         
+9     10    11    12    1  0          0.33       3         
+9     10    11    12    1  180.00     1.96       1         
+10    11    12    13    1  0          0.12       4         
+10    11    12    13    1  0          0.18       2         
+10    11    12    13    1  0          0.33       3         
+10    11    12    13    1  180.00     1.96       1         
+11    12    13    14    1  0          0.12       4         
+11    12    13    14    1  0          0.18       2         
+11    12    13    14    1  0          0.33       3         
+11    12    13    14    1  180.00     1.96       1         
+12    13    14    15    1  0          0.12       4         
+12    13    14    15    1  0          0.18       2         
+12    13    14    15    1  0          0.33       3         
+12    13    14    15    1  180.00     1.96       1         
+13    14    15    16    1  0          0.12       4         
+13    14    15    16    1  0          0.18       2         
+13    14    15    16    1  0          0.33       3         
+13    14    15    16    1  180.00     1.96       1         
+14    15    16    17    1  0          0.12       4         
+14    15    16    17    1  0          0.18       2         
+14    15    16    17    1  0          0.33       3         
+14    15    16    17    1  180.00     1.96       1         
+15    16    17    18    1  0          0.12       4         
+15    16    17    18    1  0          0.18       2         
+15    16    17    18    1  0          0.33       3         
+15    16    17    18    1  180.00     1.96       1         
+16    17    18    19    1  0          0.12       4         
+16    17    18    19    1  0          0.18       2         
+16    17    18    19    1  0          0.33       3         
+16    17    18    19    1  180.00     1.96       1         
+17    18    19    20    1  0          0.12       4         
+17    18    19    20    1  0          0.18       2         
+17    18    19    20    1  0          0.33       3         
+17    18    19    20    1  180.00     1.96       1
