+++ tests/structure_tool/3.itp
@@ -1,0 +1,11 @@
+[   moleculetype   ]
+PEO   1
+[      atoms       ]
+1     PEO   1 PEG   O1  1 0.000    45 
+2     PEO   1 PEG   O1  2 0.000    45 
+3     PEO   1 PEG   O1  3 0.000    45 
+[      bonds       ]
+1     2     1  0.330    7000   
+2     3     1  0.330    7000   
+[      angles      ]
+1     2     3     1  130.00   50
