+++ tests/structure_tool/4.itp
@@ -1,0 +1,19 @@
+[   moleculetype   ]
+PEO   1
+[      atoms       ]
+1     PEO   1 PEG   O1  1 0.000    45 
+2     PEO   1 PEG   O1  2 0.000    45 
+3     PEO   1 PEG   O1  3 0.000    45 
+4     PEO   1 PEG   O1  4 0.000    45 
+[      bonds       ]
+1     2     1  0.330    7000   
+2     3     1  0.330    7000   
+3     4     1  0.330    7000   
+[      angles      ]
+1     2     3     1  130.00   50      
+2     3     4     1  130.00   50      
+[    dihedrals     ]
+1     2     3     4     1  0          0.12       4         
+1     2     3     4     1  0          0.18       2         
+1     2     3     4     1  0          0.33       3         
+1     2     3     4     1  180.00     1.96       1
