+++ tests/structure_tool/conf.gro
