+++ tests/structure_tool/confout.gro
@@ -1,0 +1,6 @@
+PEO 27 in W
+    3
+    1PEG     O1    1   0.389   0.021   0.000  1.9655  0.6834  0.0126
+    1PEG     O1    2   0.185  -0.101   0.274 -0.0103  0.0186 -0.1201
+    1PEG     O1    3  -0.063  -0.139   0.011 -1.9552 -0.7020  0.1075
+  10.00000  10.00000  10.00000
