+++ tests/structure_tool/energy.xvg
@@ -1,0 +1,25 @@
+# This file was created Sun Dec 24 13:46:04 2017
+# Created by:
+#                  :-) GROMACS - gmx energy, VERSION 5.1.3 (-:
+# 
+# Executable:   /usr/local/gromacs/bin/gmx_mpi
+# Data prefix:  /usr/local/gromacs
+# Command line:
+#   gmx_mpi energy
+# gmx energy is part of G R O M A C S:
+#
+# Gromacs Runs On Most of All Computer Systems
+#
+@    title "GROMACS Energies"
+@    xaxis  label "Time (ps)"
+@    yaxis  label "(kJ/mol)"
+@TYPE xy
+@ view 0.15, 0.15, 0.75, 0.85
+@ legend on
+@ legend box on
+@ legend loctype view
+@ legend 0.78, 0.8
+@ legend length 2
+@ s0 legend "LJ (SR)"
+    0.000000   71.550430
+    0.030000   -2.508485
