+++ tests/structure_tool/martini_v_FG_PEG.ions.itp
@@ -1,0 +1,55 @@
+;;; IONS
+;
+; note, ions are not to be taken too seriously using a CG presentation.
+; long range electrostatic interactions are absent, and for small ions
+; the first hydration shell is considered an implicit part of the CG ion. 
+
+;;;;;; HYDRATED SODIUM ION
+;
+; same particle type would represent HYDRATED POTASSIUM
+
+[moleculetype]
+; molname 	nrexcl
+  NA+ 		1
+
+[atoms]
+;id 	type 	resnr 	residu 	atom 	cgnr 	charge
+ 1 	Qd 	1 	ION 	NA+ 	1 	1.0 
+
+
+;;;;;; HYDRATED CHLORIDE ION
+
+[moleculetype]
+; molname 	nrexcl
+  CL- 		1
+
+[atoms]
+;id 	type 	resnr 	residu 	atom 	cgnr 	charge
+ 1 	Qa 	1 	ION 	CL- 	1 	-1.0 
+
+
+;;;;;; CHOLINE ION 
+
+[moleculetype]
+; molname 	nrexcl
+  NC3+ 		1
+
+[atoms]
+;id 	type 	resnr 	residu 	atom 	cgnr 	charge
+ 1 	Q0 	1 	ION 	NC3 	1 	1.0 
+
+
+;;;;;; HYDRATED CALCIUM ION 
+;
+; Warning: this is untested, and likely not very realistic description of
+; a calcium ion. For a proper description of calcium, the solvation free
+; energy should be increased. This could be achieved in an ad-hoc manner
+; through increasing of the LJ well depth for water-calcium interactions.
+
+[moleculetype]
+; molname 	nrexcl
+  CA+ 		1
+
+[atoms]
+;id 	type 	resnr 	residu 	atom 	cgnr 	charge
+ 1 	Qd 	1 	ION 	CA+ 	1 	2.0
