+++ tests/structure_tool/out
@@ -1,0 +1,8 @@
+bonds 5.23852944873e-29
+angle 0.594642183665
+dihedral 3.83828540811
+vdw -3.42773128347
+Total Energy: 1.00519630831
+Radius of Gyration: 0.4804436853567007
+Number of rejected MC steps: 3
+
