+++ tests/structure_tool/out.gro
@@ -1,0 +1,103 @@
+Monte Carlo generated PEO
+   100     
+    1PEO    PEO    1   0.330   0.000   0.000
+    1PEO    PEO    2   0.327  -0.134   0.302
+    1PEO    PEO    3   0.619  -0.128   0.456
+    1PEO    PEO    4   0.702   0.173   0.561
+    1PEO    PEO    5   0.752   0.470   0.697
+    1PEO    PEO    6   1.003   0.544   0.897
+    1PEO    PEO    7   1.167   0.416   1.154
+    1PEO    PEO    8   1.007   0.129   1.181
+    1PEO    PEO    9   0.805  -0.038   0.981
+    1PEO    PEO   10   1.013  -0.224   0.805
+    1PEO    PEO   11   1.319  -0.291   0.700
+    1PEO    PEO   12   1.625  -0.374   0.791
+    1PEO    PEO   13   1.853  -0.344   1.028
+    1PEO    PEO   14   1.790  -0.089   1.227
+    1PEO    PEO   15   1.520  -0.160   1.402
+    1PEO    PEO   16   1.267  -0.357   1.484
+    1PEO    PEO   17   0.954  -0.413   1.396
+    1PEO    PEO   18   0.890  -0.595   1.128
+    1PEO    PEO   19   1.074  -0.732   0.891
+    1PEO    PEO   20   1.130  -0.819   0.577
+    1PEO    PEO   21   1.195  -1.126   0.477
+    1PEO    PEO   22   1.512  -1.215   0.469
+    1PEO    PEO   23   1.766  -1.212   0.681
+    1PEO    PEO   24   1.868  -1.522   0.630
+    1PEO    PEO   25   2.109  -1.513   0.404
+    1PEO    PEO   26   2.143  -1.607   0.090
+    1PEO    PEO   27   1.910  -1.804  -0.037
+    1PEO    PEO   28   1.611  -1.915   0.047
+    1PEO    PEO   29   1.326  -1.875  -0.115
+    1PEO    PEO   30   1.262  -1.555  -0.066
+    1PEO    PEO   31   1.444  -1.403   0.163
+    1PEO    PEO   32   1.750  -1.296   0.222
+    1PEO    PEO   33   1.945  -1.043   0.141
+    1PEO    PEO   34   1.831  -0.839  -0.092
+    1PEO    PEO   35   2.033  -0.757  -0.340
+    1PEO    PEO   36   2.018  -0.884  -0.644
+    1PEO    PEO   37   2.022  -0.716  -0.928
+    1PEO    PEO   38   2.239  -0.886  -1.109
+    1PEO    PEO   39   2.403  -1.117  -0.940
+    1PEO    PEO   40   2.467  -1.381  -1.129
+    1PEO    PEO   41   2.372  -1.696  -1.124
+    1PEO    PEO   42   2.185  -1.905  -1.299
+    1PEO    PEO   43   2.082  -2.047  -1.578
+    1PEO    PEO   44   2.247  -1.953  -1.848
+    1PEO    PEO   45   2.573  -1.914  -1.818
+    1PEO    PEO   46   2.872  -1.973  -1.944
+    1PEO    PEO   47   2.828  -2.217  -2.163
+    1PEO    PEO   48   2.723  -2.124  -2.462
+    1PEO    PEO   49   2.827  -2.177  -2.771
+    1PEO    PEO   50   2.704  -2.328  -3.037
+    1PEO    PEO   51   2.380  -2.279  -3.073
+    1PEO    PEO   52   2.105  -2.375  -3.228
+    1PEO    PEO   53   1.915  -2.253  -3.468
+    1PEO    PEO   54   1.796  -2.473  -3.684
+    1PEO    PEO   55   2.013  -2.615  -3.889
+    1PEO    PEO   56   2.340  -2.609  -3.930
+    1PEO    PEO   57   2.566  -2.402  -3.806
+    1PEO    PEO   58   2.615  -2.153  -3.595
+    1PEO    PEO   59   2.435  -1.879  -3.638
+    1PEO    PEO   60   2.573  -1.584  -3.588
+    1PEO    PEO   61   2.636  -1.321  -3.398
+    1PEO    PEO   62   2.619  -1.039  -3.228
+    1PEO    PEO   63   2.470  -1.002  -2.936
+    1PEO    PEO   64   2.598  -1.153  -2.672
+    1PEO    PEO   65   2.774  -0.994  -2.443
+    1PEO    PEO   66   3.019  -1.154  -2.291
+    1PEO    PEO   67   3.110  -1.331  -2.027
+    1PEO    PEO   68   3.238  -1.140  -1.790
+    1PEO    PEO   69   3.120  -0.999  -1.517
+    1PEO    PEO   70   2.959  -0.716  -1.573
+    1PEO    PEO   71   2.919  -0.389  -1.553
+    1PEO    PEO   72   3.149  -0.291  -1.339
+    1PEO    PEO   73   3.460  -0.369  -1.258
+    1PEO    PEO   74   3.558  -0.445  -0.953
+    1PEO    PEO   75   3.833  -0.332  -0.811
+    1PEO    PEO   76   4.071  -0.548  -0.885
+    1PEO    PEO   77   4.077  -0.670  -1.191
+    1PEO    PEO   78   3.861  -0.905  -1.276
+    1PEO    PEO   79   3.552  -1.019  -1.262
+    1PEO    PEO   80   3.405  -1.113  -0.982
+    1PEO    PEO   81   3.652  -1.028  -0.781
+    1PEO    PEO   82   3.964  -1.093  -0.695
+    1PEO    PEO   83   4.191  -1.076  -0.456
+    1PEO    PEO   84   4.302  -1.253  -0.200
+    1PEO    PEO   85   4.510  -1.269   0.055
+    1PEO    PEO   86   4.784  -1.090   0.093
+    1PEO    PEO   87   4.723  -0.767   0.119
+    1PEO    PEO   88   4.671  -0.694  -0.198
+    1PEO    PEO   89   4.886  -0.463  -0.294
+    1PEO    PEO   90   4.939  -0.138  -0.305
+    1PEO    PEO   91   5.185   0.049  -0.190
+    1PEO    PEO   92   5.258   0.355  -0.090
+    1PEO    PEO   93   5.077   0.624  -0.026
+    1PEO    PEO   94   4.764   0.539   0.033
+    1PEO    PEO   95   4.586   0.574  -0.242
+    1PEO    PEO   96   4.715   0.398  -0.490
+    1PEO    PEO   97   4.976   0.470  -0.679
+    1PEO    PEO   98   5.081   0.747  -0.824
+    1PEO    PEO   99   4.844   0.908  -0.988
+    1PEO    PEO  100   4.758   1.226  -0.992
+  10.00000 10.00000 10.00000
